Refactor admin and dashboard structure
