+++ frontend/src/features/devices/transport/fetchDataSourceCsv.ts
@@ -1,0 +1,7 @@
+export const fetchDataSourceCsv = async (sourceUrl: string) => {
+  const response = await fetch(sourceUrl);
+  if (!response.ok) {
+    throw new Error("Failed to download data");
+  }
+  return response.text();
+};
