+++ frontend/src/features/devices/types.ts
@@ -1,0 +1,25 @@
+export interface DeviceInfo {
+  id: string;
+  name: string;
+  type: "Camera" | "Sensor" | "Gateway";
+  status: "online" | "offline" | "maintenance";
+  lastSeen: string;
+  dataSource?: string;
+  location?: string;
+  recordCount?: number;
+}
+
+export interface DataSource {
+  id: string;
+  title: string;
+  url: string;
+  type: string;
+  active?: boolean;
+}
+
+export interface DeviceUser {
+  role: "admin" | "client";
+  name: string;
+  csv_url?: string;
+  data_sources?: DataSource[];
+}
